vps/pyproject.toml
[project]
name = "vps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=5.3.0",
    "fastapi>=0.127.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
    "google-ai-generativelanguage>=0.6.15",
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.5",
    "python-multipart>=0.0.21",
    "pymongo>=4.15.5",
]

[project.scripts]
vps = "vps.app:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
